RealTimeDataMonitorDemo/pyproject.toml
[tool.poetry]
name = "realtime-data-monitor-demo"
version = "0.1.0"
description = "FastAPI + gRPC + asyncio queue 的实时数据监控 demo"
authors = ["Your Name <you@example.com>"]
license = "MIT"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111"
uvicorn = { extras = ["standard"], version = "^0.30" }
grpcio = "^1.73"
grpcio-tools = "^1.73"
pydantic = "^2.7"
pytest = "^8.2"
pytest-asyncio = "^0.23"
redis = "^6.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
